[Feature branch] Add FusedMOE for DeepSeek (#3533)
* init PR

* add prepack

* add woq_int8

* Revert "Update decoder.py"

This reverts commit 9d7bf17b044498de883cc34c23b66f1399b0c628.

* reduce mem

* code refinement and API init

* code refine

* [Feature branch] Add avx GEMM in FusedMOE for m_size < 4 (#3553)

* avoid dup define pack

* int8 load fusedmoe support

* refine quant/dequant

* minor fix for dtype

+++ intel_extension_for_pytorch/transformers/models/reference/models.py
@@ -5845,9 +5845,7 @@ def JambaForCausalLM_forward(
 
 def Deepseek_MoEGate_forward(self, hidden_states):
     # compute gating score
-    logits = torch.nn.functional.linear(
-        hidden_states.type(torch.float32), self.weight.type(torch.float32), None
-    )
+    logits = torch.nn.functional.linear(hidden_states, self.weight, None)
 
     if self.scoring_func == "softmax":
         scores = logits.softmax(dim=-1, dtype=hidden_states.dtype)
